requirements.txt
absl-py==0.12.0
appdirs==1.4.4
argcomplete==1.12.0
arrow==0.17.0
astunparse==1.6.3
Babel==2.9.1
Beaker==1.10.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
blivet==3.3.3
blivet-gui==2.3.0
Brlapi==0.8.2
cachetools==4.2.2
cffi==1.14.5
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
conda==4.10.1
conda-package-handling==1.7.2
cookiecutter==1.7.3
cryptography==3.4.6
cupshelpers==1.0
cytoolz==0.11.0
dasbus==1.4
dbus-python==1.2.18
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
filelock==3.0.12
flatbuffers==1.12
fros==1.1
frozendict==1.2
future==0.18.2
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gpg==1.15.1
grpcio==1.34.1
h5py==3.1.0
humanize==0.5.1
idna==2.10
Jinja2==2.11.3
jinja2-time==0.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
langtable==0.0.56
libcomps==0.1.18
lxml==4.6.3
Mako==1.1.4.dev0
Markdown==3.3.4
MarkupSafe==1.1.1
nftables==0.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==8.1.2
ply==3.11
poyo==0.5.0
productmd==1.33
protobuf==3.17.0
ptyprocess==0.6.0
pwquality==1.4.4
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.1
PyGObject==3.40.1
pykickstart==3.32
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyparted==3.11.7
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.1
python-meh==0.50
python-slugify==4.0.1
pytz==2021.1
pyudev==0.22.0
pyxdg==0.27
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.0
rpm==4.16.1.3
rsa==4.7.2
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.2
sepolicy==3.2
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.8
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.1
soupsieve==2.2
systemd-python==234
Tempita==0.5.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
text-unidecode==1.3
toolz==0.11.1
tqdm==4.62.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.4.0
Werkzeug==2.0.0
wrapt==1.12.1
